+++ .rubocop_todo.yml
@@ -810,12 +810,6 @@ Rails/ApplicationJob:
   Exclude:
     - 'app/jobs/report_job.rb'
 
-# Offense count: 1
-# This cop supports unsafe autocorrection (--autocorrect-all).
-Rails/ApplicationMailer:
-  Exclude:
-    - 'app/mailers/spree/base_mailer.rb'
-
 # Offense count: 5
 # This cop supports unsafe autocorrection (--autocorrect-all).
 # Configuration parameters: NilOrEmpty, NotPresent, UnlessPresent.
