+++ app/mailers/spree/user_mailer.rb
@@ -3,7 +3,7 @@
 # This mailer is configured to be the Devise mailer
 # Some methods here override Devise::Mailer methods
 module Spree
-  class UserMailer < BaseMailer
+  class UserMailer < ApplicationMailer
     include I18nHelper
 
     helper MailerHelper
